@@ -49,9 +49,9 @@
     "lucide-react": "^0.462.0",
     "netlify-cms-app": "^2.15.72",
     "next-themes": "^0.3.0",
-    "react": "^17.0.2",
+    "react": "^18.3.1",
     "react-day-picker": "^8.10.1",
-    "react-dom": "^17.0.2",
+    "react-dom": "^18.3.1",
     "react-hook-form": "^7.53.0",
     "react-resizable-panels": "^2.1.3",
     "react-router-dom": "^6.26.2",
