@@ -1,6 +1,5 @@
-
-import ReactDOM from 'react-dom'
+import { createRoot } from 'react-dom/client'
 import App from './App.tsx'
 import './index.css'
 
-ReactDOM.render(<App />, document.getElementById("root"));
+createRoot(document.getElementById("root")!).render(<App />);
